Updates to Post-method and added support for different api url.
Due to incompatibility between Ceobackend and the Post-method using HttpURlConnection, I've added a new Post Method that uses the HTTPClient-method, that works with both CeoBackend and SkatBackend.
Please note that the HttpClient is set to Version 1_1

Updated the way the app is run, so to better support running with and changing between different api URL.

With the sucess of the HTTPClient in POST, I should consider doing the same woth for the GET-method.