Changes and Updates to POST- and GET-method + Discovery of new bug
Post: Moved old HTTPURLConnection method below the new httpclient method, though the old is still in the code for comparrison sake. The App uses the httpclient when run

Get: Similiar to the Post, I've added a Get-,ethod that uses the httpclient and made teh app use that, instead of the old httpurlconnection.

Moved the static HttpClient to its own separat class, since its used by both GET and POST

Bugs: Discovered a calculation-bug, when calculating a score with lots of spares! A triple spare-game is correctly calculated, but something goes horribly wrong in games with only spares, and games with lots of spares in general with 2-3 spares in a row. Have done a few test to show this bug, pretty clear some calcs are off. Need to fix this!