Bug fixing and readme-update
Spare-points bug fixing: Fixed a bug that resultet in incorrect calculations for spares, if there were multiple spares in succession in a large game.

Normal-points bug fixing: Fixed a bug that resultet in incorrect calculations for a normal score in the 9th-frame, if the player successfully got 3 strikes in the tenth- and eleventh-bonus frame.

Readme: Minor readme update regarding the apiConn.getCeoBackend()